pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bridgesql"
version = "0.1.0"
description = "Natural language interface for your database — MCP server with semantic catalog"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Grad Project" }
]
keywords = ["nl2sql", "semantic", "database", "llm", "mcp", "bridgesql"]

dependencies = [
    # Database
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "cryptography>=42.0.0",  # MySQL SSL support
    
    # LLM - Gemini
    "google-generativeai>=0.8.0",
    
    # Vector Store
    "chromadb>=0.5.0",
    
    # Data Processing
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pandas>=2.0.0",
    
    # SQL Parsing & Validation
    "sqlparse>=0.5.0",
    "sqlglot>=25.0.0",
    
    # MCP Server
    "mcp>=1.0.0",
    
    # Utilities
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "sentence-transformers>=3.0.0",
    
    # CLI prompts
    "prompt_toolkit>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]
slack = [
    "slack-bolt>=1.20.0",
]

[project.scripts]
sqe = "semantic_engine.cli:main"
sqe-mcp = "semantic_engine.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/semantic_engine"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
